Revert "Suspending a cluster now sends the right message when in progress (#364)"
This reverts commit e59064715d4e6a139f850dbe239ca6e508fb35e0.
