+++ crate/operator/expand_volume.py
@@ -35,10 +35,6 @@ from crate.operator.webhooks import (
     WebhookStatus,
 )
 
-EXPAND_REPLICAS_IN_PROGRESS_MSG = (
-    "Suspending cluster and waiting for Persistent Volume Claim(s) to be resized."
-)
-
 
 async def expand_volume(
     core: CoreV1Api,
@@ -63,7 +59,8 @@ async def expand_volume(
     await send_operation_progress_notification(
         namespace=namespace,
         name=name,
-        message=EXPAND_REPLICAS_IN_PROGRESS_MSG,
+        message="Suspending cluster and waiting for Persistent "
+        "Volume Claim(s) to be resized.",
         logger=logger,
         status=WebhookStatus.IN_PROGRESS,
         operation=WebhookOperation.UPDATE,
